backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
anthropic==0.49.0
openai==1.75.0
apscheduler==3.10.4
